BeepBoop/pyproject.toml
[project]
name = "beepboop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.8.0",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
]
